Fixing threading on control_module:
previously the control_module would start a thread on __init__, but since _running wasn't set to true the thread would exit immediately. The thread should sync values when it finished (ie. call the various to_COPY methods, without with tests would sporadically fail), but initing the thread causes that to fail and be caught on the lock if the controller is never started.

Changed so that thread is only started with `start`, retaining previous commit that calls to_COPY methods at end of _start_mainloop